src/backend/trade_data_ws/requirements.txt
pulsar-client==3.5.0
websockets==14.2